# gateway-config/smarthealth-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "description": "Protected route: requires a JWT token with ADMIN role. Use the \"pnpm run seed\" script to create the first administrator.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[ADMIN ONLY] Create a Doctor or Admin account",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Log in and obtain a JWT Token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/users": {
      "get": {
        "operationId": "AuthController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[ADMIN ONLY] List all users",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/users/{id}": {
      "patch": {
        "operationId": "AuthController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[ADMIN ONLY] Update a user",
        "tags": [
          "auth"
        ]
      },
      "delete": {
        "operationId": "AuthController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[ADMIN ONLY] Delete a user",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartHealth API",
    "description": "Interactive OpenAPI documentation for the SmartHealth API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "doctor@smarthealth.com"
          },
          "password": {
            "type": "string",
            "example": "securepassword123"
          },
          "role": {
            "type": "string",
            "example": "DOCTOR",
            "description": "Role of the new account (DOCTOR or ADMIN)",
            "enum": [
              "DOCTOR",
              "ADMIN"
            ],
            "default": "DOCTOR"
          },
          "firstName": {
            "type": "string",
            "example": "Jean"
          },
          "lastName": {
            "type": "string",
            "example": "Rakoto"
          },
          "specialty": {
            "type": "string",
            "example": "Cardiologue"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "doctor@smarthealth.com"
          },
          "password": {
            "type": "string",
            "example": "securepassword123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}